+++ src/backend/core/api/serializers.py
@@ -170,7 +170,6 @@ class TeamSerializer(serializers.ModelSerializer):
     """Serialize teams."""
 
     abilities = serializers.SerializerMethodField(read_only=True)
-    slug = serializers.SerializerMethodField()
 
     class Meta:
         model = models.Team
@@ -179,7 +178,6 @@ class TeamSerializer(serializers.ModelSerializer):
             "name",
             "accesses",
             "abilities",
-            "slug",
             "created_at",
             "updated_at",
         ]
@@ -187,7 +185,6 @@ class TeamSerializer(serializers.ModelSerializer):
             "id",
             "accesses",
             "abilities",
-            "slug",
             "created_at",
             "updated_at",
         ]
@@ -198,10 +195,6 @@ class TeamSerializer(serializers.ModelSerializer):
         if request:
             return team.get_abilities(request.user)
         return {}
-
-    def get_slug(self, instance):
-        """Return slug from the team's name."""
-        return instance.get_slug()
 
 
 class InvitationSerializer(serializers.ModelSerializer):